requirements.txt
fpdf
langchain-community
langchain_groq
streamlit
python-dotenv
python-docx